requirements.txt
streamlit==1.31.0
requests==2.31.0
pandas==2.1.4
python-dotenv==1.0.0
mysql-connector-python==8.2.0
mlflow==2.9.2
boto3==1.34.19
plotly==5.18.0
openpyxl==3.1.2
pytest==7.4.3